[GTK][WPE] Centralize GLFence creation boilerplate in SkiaUtilities https://bugs.webkit.org/show_bug.cgi?id=311180
Reviewed by Adrian Perez de Castro.

Move the duplicated "flush + submit + create fence, fallback to sync submit"
pattern into three new SkiaUtilities functions: flushAndSubmitSurfaceWithFence,
flushAndSubmitImageWithFence, and flushAndSubmitWithFence.

This eliminates repeated boilerplate across GraphicsContextSkia,
ImageBufferSkiaAcceleratedBackend, CoordinatedTileBuffer, and SkiaPaintingEngine.

Also removes GraphicsContextSkia::createAcceleratedRenderingFence since all
call sites already ensure the GL context is current -- simplifying the
desing.

Covered by existing tests.

Canonical link: https://commits.webkit.org/310333@main
