+++ Source/WebCore/platform/graphics/skia/GraphicsContextSkia.cpp
@@ -341,7 +341,7 @@ void GraphicsContextSkia::drawNativeImage(const NativeImage& nativeImage, const 
             if (!image->isValid(grContext->asRecorder())) {
                 // Ensure any pending GPU operations on the source image are complete before
                 // accessing its backend texture for rewrapping.
-                if (auto fence = createAcceleratedRenderingFence(nativeImage.platformImage(), nativeImage.grContext()))
+                if (auto fence = SkiaUtilities::flushAndSubmitImageWithFence(nativeImage.grContext(), nativeImage.platformImage()))
                     fence->serverWait();
 
                 imageInThisThread = SkiaUtilities::rewrapImageForContext(grContext, *image);
@@ -1292,48 +1292,6 @@ void GraphicsContextSkia::replayStateOnCanvas(SkCanvas& canvas) const
     canvas.setMatrix(m_canvas.getTotalMatrix());
 }
 
-static std::unique_ptr<GLFence> createFenceAfterFlush(GrDirectContext* grContext)
-{
-    auto& glDisplay = PlatformDisplay::sharedDisplay().glDisplay();
-    if (GLFence::isSupported(glDisplay)) {
-        grContext->submit(GrSyncCpu::kNo);
-
-        if (auto fence = GLFence::create(glDisplay))
-            return fence;
-    }
-
-    grContext->submit(GrSyncCpu::kYes);
-    return nullptr;
-}
-
-std::unique_ptr<GLFence> GraphicsContextSkia::createAcceleratedRenderingFence(SkSurface* surface)
-{
-    auto* glContext = PlatformDisplay::sharedDisplay().skiaGLContext();
-    if (!glContext || !glContext->makeContextCurrent())
-        return nullptr;
-
-    auto* recordingContext = surface->recordingContext();
-    auto* grContext = recordingContext ? recordingContext->asDirectContext() : nullptr;
-    if (!grContext)
-        return nullptr;
-
-    grContext->flush(surface);
-    return createFenceAfterFlush(grContext);
-}
-
-std::unique_ptr<GLFence> GraphicsContextSkia::createAcceleratedRenderingFence(const sk_sp<SkImage>& image, GrDirectContext* grContext)
-{
-    auto* glContext = PlatformDisplay::sharedDisplay().skiaGLContext();
-    if (!glContext || !glContext->makeContextCurrent())
-        return nullptr;
-
-    if (!grContext)
-        return nullptr;
-
-    grContext->flush(image);
-    return createFenceAfterFlush(grContext);
-}
-
 void GraphicsContextSkia::trackAcceleratedRenderingFenceIfNeeded(const sk_sp<SkImage>& image, GrDirectContext* grContext)
 {
     if (m_contextMode != ContextMode::RecordingMode)
@@ -1342,7 +1300,7 @@ void GraphicsContextSkia::trackAcceleratedRenderingFenceIfNeeded(const sk_sp<SkI
     if (!image || !image->isTextureBacked())
         return;
 
-    if (auto fence = createAcceleratedRenderingFence(image, grContext))
+    if (auto fence = SkiaUtilities::flushAndSubmitImageWithFence(grContext, image))
         m_imageToFenceMap.add(image.get(), WTF::move(fence));
 }
